requirement.txt
amqp==2.2.2
ansible==2.4.2.0
Babel==0.9.6
backports.ssl-match-hostname==3.4.0.2
billiard==3.5.0.3
boto==2.45.0
celery==4.1.0
cffi==1.6.0
chardet==2.2.1
Cheetah==2.4.4
click==6.7
configobj==4.7.2
cryptography==1.7.2
decorator==3.4.0
DenyHosts==2.6
enum34==1.0.4
Flask==0.12.2
gunicorn==19.7.1
honcho==1.0.1
httplib2==0.9.2
idna==2.4
iniparse==0.4
ipaddress==1.0.16
IPy==0.75
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.0
jsonpatch==1.2
jsonpointer==1.9
kombu==4.1.0
Markdown==2.4.1
MarkupSafe==0.11
paramiko==2.1.1
passlib==1.6.5
perf==0.1
Pillow==2.0.0
ply==3.4
policycoreutils-default-encoding==0.1
prettytable==0.7.2
psutil==5.4.5
pyasn1==0.1.9
pycparser==2.14
pycurl==7.19.0
Pygments==1.4
pygobject==3.14.0
pygpgme==0.3
pyliblzma==0.5.3
python-envcfg==0.2.0
pytz==2018.3
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.10
redis==2.10.6
requests==2.6.0
rsa==3.4.1
seobject==0.1
sepolicy==1.1
six==1.9.0
slip==0.4.0
slip.dbus==0.4.0
urlgrabber==3.10
urllib3==1.10.2
vine==1.1.4
virtualenv==15.2.0
Werkzeug==0.14.1
yum-metadata-parser==1.1.4
